Redesign test layout structure
