+++ bokehjs/make/tasks/lint.ts
@@ -51,18 +51,18 @@ async function eslint(dir: string): Promise<void> {
   }
 }
 
-task("eslint:test:unit", async () => await eslint(join(paths.src_dir.test, "unit")))
+task("eslint:test:framework", async () => await eslint(join(paths.src_dir.test, "framework")))
 task("eslint:test:defaults", async () => await eslint(join(paths.src_dir.test, "defaults")))
+task("eslint:test:unit", async () => await eslint(join(paths.src_dir.test, "unit")))
 task("eslint:test:integration", async () => await eslint(join(paths.src_dir.test, "integration")))
 task("eslint:test:codebase", async () => await eslint(join(paths.src_dir.test, "codebase")))
 task("eslint:test:devtools", async () => await eslint(join(paths.src_dir.test, "devtools")))
-task("eslint:test:common", async () => await eslint(paths.src_dir.test))
 
 task("eslint:make", async () => await eslint(paths.make_dir))
 task("eslint:lib", async () => await eslint(paths.src_dir.lib))
 task("eslint:compiler", async () => await eslint(paths.src_dir.compiler))
 task("eslint:server", async () => await eslint(paths.src_dir.server))
-task("eslint:test", ["eslint:test:unit", "eslint:test:defaults", "eslint:test:integration", "eslint:test:codebase", "eslint:test:devtools", "eslint:test:common"])
+task("eslint:test", ["eslint:test:framework", "eslint:test:defaults", "eslint:test:unit", "eslint:test:integration", "eslint:test:codebase", "eslint:test:devtools"])
 task("eslint:examples", async () => await eslint(paths.src_dir.examples))
 
 task("eslint", ["eslint:make", "eslint:lib", "eslint:compiler", "eslint:server", "eslint:test", "eslint:examples"])
