+++ bokehjs/test/framework/interactive.ts
