+++ bokehjs/test/unit/models/annotations/legend_item.ts
@@ -1,4 +1,4 @@
-import {expect} from "assertions"
+import {expect} from "#framework/assertions"
 import * as sinon from "sinon"
 
 import {ColumnDataSource} from "@bokehjs/models/sources/column_data_source"
